+++ internal/provider/data_account_test.go
@@ -2,11 +2,13 @@ package provider
 
 import (
 	"fmt"
+	"github.com/hashicorp/terraform-plugin-testing/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 	"testing"
 )
 
 func TestAccDataAccount_default(t *testing.T) {
+	name := acctest.RandomWithPrefix("tfacc")
 	resource.ParallelTest(t, resource.TestCase{
 		PreCheck: func() {
 			preCheck(t)
@@ -15,7 +17,7 @@ func TestAccDataAccount_default(t *testing.T) {
 		// TODO: CheckDestroy: ,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccDataAccountConfig("tfusertest", "secure_1234"),
+				Config: testAccDataAccountConfig(name, "secure_1234"),
 				Check:  resource.ComposeTestCheckFunc(),
 			},
 		},
@@ -23,6 +25,9 @@ func TestAccDataAccount_default(t *testing.T) {
 }
 
 func TestAccDataAccount_mac(t *testing.T) {
+	mac, unallocateMac := allocateTestMac(t)
+	defer unallocateMac()
+
 	resource.ParallelTest(t, resource.TestCase{
 		PreCheck: func() {
 			preCheck(t)
@@ -31,7 +36,7 @@ func TestAccDataAccount_mac(t *testing.T) {
 		// TODO: CheckDestroy: ,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccDataMacAccountConfig("00B0D06FC226"),
+				Config: testAccDataMacAccountConfig(mac),
 				Check:  resource.ComposeTestCheckFunc(),
 			},
 		},
@@ -41,31 +46,31 @@ func TestAccDataAccount_mac(t *testing.T) {
 func testAccDataAccountConfig(name, password string) string {
 	return fmt.Sprintf(`
 resource "unifi_account" "test" {
-	name = "%s"
-	password = "%s"
+	name = "%[1]s"
+	password = "%[2]s"
 }
 
 data "unifi_account" "test" {
-	name = "%s"
+	name = "%[1]s"
 depends_on = [
     unifi_account.test
   ]
 }
-`, name, password, name)
+`, name, password)
 }
 
 func testAccDataMacAccountConfig(mac string) string {
 	return fmt.Sprintf(`
 resource "unifi_account" "test" {
-	name = "%s"
-	password = "%s"
+	name = "%[1]s"
+	password = "%[1]s"
 }
 
 data "unifi_account" "test" {
-	name = "%s"
+	name = "%[1]s"
 depends_on = [
     unifi_account.test
   ]
 }
-`, mac, mac, mac)
+`, mac)
 }
