+++ internal/provider/resource_user_test.go
@@ -6,11 +6,9 @@ import (
 	"net"
 	"regexp"
 	"strings"
-	"sync"
 	"testing"
 
 	"github.com/apparentlymart/go-cidr/cidr"
-	mapset "github.com/deckarep/golang-set/v2"
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 	"github.com/hashicorp/terraform-plugin-testing/terraform"
 	"github.com/paultyng/go-unifi/unifi"
@@ -18,36 +16,6 @@ import (
 
 func userImportStep(name string) resource.TestStep {
 	return importStep(name, "allow_existing", "skip_forget_on_destroy")
-}
-
-var (
-	macInit sync.Once
-	macPool mapset.Set[*net.HardwareAddr] = mapset.NewSet[*net.HardwareAddr]()
-)
-
-func allocateTestMac(t *testing.T) (string, func()) {
-	macInit.Do(func() {
-		// for test MAC addresses, see https://tools.ietf.org/html/rfc7042#section-2.1.
-		for i := 0; i < 256; i++ {
-			mac := net.HardwareAddr{0x00, 0x00, 0x5e, 0x00, 0x53, byte(i)}
-			if ok := macPool.Add(&mac); !ok {
-				t.Fatal("Failed to add MAC to pool")
-			}
-		}
-	})
-
-	mac, ok := macPool.Pop()
-	if mac == nil || !ok {
-		t.Fatal("Unable to allocate test MAC")
-	}
-
-	unallocate := func() {
-		if ok := macPool.Add(mac); !ok {
-			t.Fatal("Failed to add MAC to pool")
-		}
-	}
-
-	return mac.String(), unallocate
 }
 
 func TestAccUser_basic(t *testing.T) {
